--- src/main/java/com/c9pay/authservice/dto/CertificateResponse.java
@@ -1,12 +1,0 @@
-package com.c9pay.authservice.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class CertificateResponse {
-    private String certificate;
-}
