--- src/main/java/com/c9pay/authservice/dto/SerialNumberResponse.java
@@ -1,14 +1,0 @@
-package com.c9pay.authservice.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.UUID;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class SerialNumberResponse {
-    private UUID serialNumber;
-}
