+++ src/main/java/com/c9pay/authservice/web/dto/CertificateResponse.java
@@ -1,4 +1,4 @@
-package com.c9pay.authservice.dto;
+package com.c9pay.authservice.web.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,6 +7,6 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class PublicKeyResponse {
-    private String publicKey;
+public class CertificateResponse {
+    private String certificate;
 }
