+++ src/main/java/com/c9pay/authservice/web/dto/PublicKeyResponse.java
@@ -1,4 +1,4 @@
-package com.c9pay.authservice.dto;
+package com.c9pay.authservice.web.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,7 +7,6 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class ServiceInfo {
-    private String serviceName;
-    private String RecognitionEndpoint;
+public class PublicKeyResponse {
+    private String publicKey;
 }
