+++ src/main/java/com/c9pay/authservice/web/dto/SerialNumberResponse.java
@@ -1,13 +1,14 @@
-package com.c9pay.authservice.dto;
+package com.c9pay.authservice.web.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.util.UUID;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class CertificateForm {
-    private String publicKey;
-    private ServiceInfo serviceInfo;
+public class SerialNumberResponse {
+    private UUID serialNumber;
 }
